% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-functions.R
\name{create_multicolour_stl}
\alias{create_multicolour_stl}
\title{Create multi-colour}
\usage{
create_multicolour_stl(df, elev, inverse = FALSE)
}
\arguments{
\item{df}{Sf polygon to rasterize}

\item{elev}{The elevation raster}

\item{inverse}{Inverse the mask TRUE/FALSE}
}
\value{
Matrix
}
\description{
Create multi-colour
}
