% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livestock-functions.R
\name{get_stock_unit}
\alias{get_stock_unit}
\title{Nearest stock unit equivalent based on year}
\usage{
get_stock_unit(animal, year, stock_units = stock_units)
}
\arguments{
\item{animal}{animal type (e.g. "sheep", "beer", "dairy", "deer")}

\item{year}{year}

\item{stock_units}{stock units data frame}
}
\value{
Nearest stock unit equivalent for the given animal and year
}
\description{
\code{get_stock_unit} returns the nearest stock unit equivalent for the given animal and year.

The function uses the stock_units data frame to find the nearest year and corresponding stock unit equivalent.
The stock_units data frame contains the stock unit equivalents for different animals
for the years 1980, 1985, 2000, and 2017. This can be updated as needed, or held constant.

Follows the methods of Snelder et al (2021).
}
\examples{
get_stock_unit("Sheep", 1990)

library(dplyr)

test_df <- tibble(
  year = rep(c(1970, 1985, 1990, 2005), each = 3),
  animal = rep(c("Sheep", "Beef cattle", "Dairy cattle"), times = 4),
  count = c(3000, 5000, 7000, 8000, 10000, 6000, 7000, 9000, 11000, 5000, 6000, 7000)
)

test_df |>
  rowwise() |>
  mutate(stock_unit_equivalent = count * get_stock_unit(animal, year)) |>
  ungroup()
}
\references{
Snelder TH, Fraser C, Larned ST, Monaghan R, De Malmanche S, Whitehead AL. Attribution of river water-quality trends to agricultural land use and climate variability in New Zealand. Marine and Freshwater Research. 2021;73(1):1-19. doi:10.1071/mf21086
}
