% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-functions.R
\name{import_clip_raster}
\alias{import_clip_raster}
\title{Import folder of tifs and clip by polygon}
\usage{
import_clip_raster(tif_folder, polygon_spatvector)
}
\arguments{
\item{tif_folder}{Folder where your raster files are stored}

\item{polygon_spatvector}{Polygon to clip the rasters to}
}
\value{
A clipped raster
}
\description{
Import folder of tifs and clip by polygon
}
