% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcdb5_chch}
\alias{lcdb5_chch}
\title{LCDB5 data cropped to Christchurch}
\format{
An SF data object with 2,232 features and 23 fields.
NZTM2000 projection. CRS: 2193.
Clipped to bbox: xmin = 1543016, xmax = 1587730, ymin = 5169087, ymax = 5190086 .
}
\source{
\url{https://lris.scinfo.org.nz/layer/104400-lcdb-v50-land-cover-database-version-50-mainland-new-zealand/}
}
\usage{
lcdb5_chch
}
\description{
A subset of Land Cover Database version 5 (LCDB5) data cropped to the Christchurch area.
}
\keyword{datasets}
