% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec-functions.R
\name{load.rivers}
\alias{load.rivers}
\title{Load REC2.5 river data}
\usage{
load.rivers(
  df = "data/REC2_geodata_version_5/nzRec2_v5.gdb",
  layer = "riverlines"
)
}
\arguments{
\item{df}{REC geodatabase file}

\item{layer}{Which layer to load}
}
\value{
An sf object with the REC river data
}
\description{
Load REC2.5 river data
}
