% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec-functions.R
\name{load.ws}
\alias{load.ws}
\title{Load REC2.5 watershed data}
\usage{
load.ws(df = "data/REC2_geodata_version_5/nzRec2_v5.gdb", layer = "rec2ws")
}
\arguments{
\item{df}{REC geodatabase file}

\item{layer}{Which layer to load}
}
\value{
An sf object with the REC watershed data
}
\description{
Load REC2.5 watershed data
}
