% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcdb-functions.R
\name{load_lcdb_symbology}
\alias{load_lcdb_symbology}
\title{Load LCDB symbology}
\usage{
load_lcdb_symbology()
}
\value{
A data frame with the symbology for the LCDB classes
}
\description{
Load LCDB symbology
}
\examples{
load_lcdb_symbology()
}
