% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcdb-functions.R
\name{load_lcdb_vector}
\alias{load_lcdb_vector}
\title{Load named vector of LCDB colours}
\usage{
load_lcdb_vector()
}
\value{
A named vector of LCDB colours
}
\description{
Load named vector of LCDB colours
}
