% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-functions.R
\name{rect_around_point}
\alias{rect_around_point}
\title{Create a Rectangle around a Point}
\usage{
rect_around_point(center, xsize, ysize, crs = 2193)
}
\arguments{
\item{center}{A numeric vector of length 2 specifying the x and y coordinates of the center of the rectangle (e.g., c(x, y)).}

\item{xsize}{Numeric, the total width of the rectangle (in the same units as the coordinate reference system, typically meters).}

\item{ysize}{Numeric, the total height of the rectangle (in the same units as the coordinate reference system, typically meters).}

\item{crs}{Numeric, the coordinate reference system (CRS) of the output geometry. Default is EPSG:2193 (NZTM2000).}
}
\value{
An \code{sf} object representing the rectangle polygon.
}
\description{
This function creates a rectangle of specified width and height, centered on a given point.
}
\examples{
\dontrun{
# Example usage: Create a rectangle with width 1000 meters and height 500 meters centered at (1577107, 5173732) in NZTM2000
rectangle_sf <- rect_around_point(center = c(1577107, 5173732), xsize = 1000, ysize = 500)

# Plot the rectangle
plot(rectangle_sf)
}

}
