% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-functions.R
\name{st_create_hexagon}
\alias{st_create_hexagon}
\title{Function to create a hexagon with a given radius, center, and rotation}
\usage{
st_create_hexagon(center, radius, rotation = 0, crs = 2193)
}
\arguments{
\item{center}{coordinates of the center of the hexagon}

\item{radius}{radius of the hexagon}

\item{rotation}{rotate corners}

\item{crs}{crs for sf object}
}
\value{
A hexagon sf object
}
\description{
Function to create a hexagon with a given radius, center, and rotation
}
\examples{
\dontrun{
hexagon_sf <- create_hexagon(center = c(1600000, 5500000), radius = 100000, rotation = pi/8, crs = 2193)
}

}
