% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf-functions.R
\name{st_crop_bbox}
\alias{st_crop_bbox}
\title{Crop sf object by bbox}
\usage{
st_crop_bbox(x, bbox)
}
\arguments{
\item{x}{sf object to crop}

\item{bbox}{bbox object to crop by}
}
\value{
A smaller sf object
}
\description{
Crop sf object by bbox
}
