% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec-functions.R
\name{trace_upstream}
\alias{trace_upstream}
\title{Trace upstream}
\usage{
trace_upstream(df, start_segment, max_distance = Inf, current_distance = 0)
}
\arguments{
\item{df}{REC sf object}

\item{start_segment}{Starting segment}

\item{max_distance}{Maximum distance to trace upstream, set to Inf to goto top of catchment}

\item{current_distance}{Set to 0}
}
\value{
A list of segments which are upstream of the starting segment
}
\description{
Trace upstream
}
